requirements.txt
flask
requests
pymysql
tencentcloud-sdk-python

